Split off `Internal.Cardano.Write.Eras`
