+++ lib/balance-tx/lib/internal/Internal/Cardano/Write/Tx.hs
@@ -8,10 +8,8 @@
 {-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilyDependencies #-}
-{-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
 
@@ -28,40 +26,14 @@
 -- wallet migration.
 module Internal.Cardano.Write.Tx
     (
-    -- * Eras
-      BabbageEra
-    , ConwayEra
-
-    -- ** RecentEra
-    , RecentEra (..)
-    , IsRecentEra (..)
-    , CardanoApiEra
-    , toRecentEra
-    , fromRecentEra
-    , MaybeInRecentEra (..)
-    , toRecentEraGADT
-    , LatestLedgerEra
-    , RecentEraConstraints
-    , allRecentEras
-
     -- ** Key witness counts
-    , KeyWitnessCounts (..)
+      KeyWitnessCounts (..)
 
     -- ** Helpers for cardano-api compatibility
-    , cardanoEra
-    , shelleyBasedEra
-    , CardanoApi.ShelleyLedgerEra
-    , cardanoEraFromRecentEra
-    , shelleyBasedEraFromRecentEra
     , fromCardanoApiTx
     , toCardanoApiUTxO
     , fromCardanoApiUTxO
     , toCardanoApiTx
-
-    -- ** Existential wrapper
-    , AnyRecentEra (..)
-    , toAnyCardanoEra
-    , fromAnyCardanoEra
 
     -- ** Misc
     , StandardCrypto
@@ -71,6 +43,7 @@ module Internal.Cardano.Write.Tx
     -- * PParams
     , PParams
     , PParamsInAnyRecentEra (..)
+    , toRecentEraGADT
     , FeePerByte (..)
     , getFeePerByte
     , feeOfBytes
@@ -155,30 +128,17 @@ import Cardano.Crypto.Hash
     ( Hash (UnsafeHash)
     )
 import Cardano.Ledger.Allegra.Scripts
-    ( AllegraEraScript
-    , Timelock
-    , translateTimelock
-    )
-import Cardano.Ledger.Alonzo.Plutus.Context
-    ( EraPlutusContext
+    ( translateTimelock
     )
 import Cardano.Ledger.Alonzo.Scripts
-    ( AlonzoEraScript
-    , AlonzoScript (..)
-    )
-import Cardano.Ledger.Alonzo.TxWits
-    ( AlonzoTxWits
-    )
-import Cardano.Ledger.Alonzo.UTxO
-    ( AlonzoScriptsNeeded
+    ( AlonzoScript (..)
     )
 import Cardano.Ledger.Api
     ( coinTxOutL
     , upgradeTxOut
     )
 import Cardano.Ledger.Api.UTxO
-    ( EraUTxO (ScriptsNeeded)
-    , UTxO (..)
+    ( UTxO (..)
     )
 import Cardano.Ledger.Babbage.TxBody
     ( BabbageTxOut (..)
@@ -234,9 +194,6 @@ import Data.ByteString.Short
 import Data.Coerce
     ( coerce
     )
-import Data.Function
-    ( on
-    )
 import Data.Generics.Internal.VL.Lens
     ( over
     , (^.)
@@ -247,25 +204,18 @@ import Data.IntCast
     ( intCast
     , intCastMaybe
     )
-import Data.Kind
-    ( Type
-    )
-import Data.Maybe
-    ( fromMaybe
-    , isJust
-    )
-import Data.Set
-    ( Set
-    )
-import Data.Type.Equality
-    ( TestEquality (testEquality)
-    , (:~:) (Refl)
-    )
-import Data.Typeable
-    ( Typeable
-    )
 import GHC.Stack
     ( HasCallStack
+    )
+import Internal.Cardano.Write.Eras
+    ( BabbageEra
+    , CardanoApiEra
+    , ConwayEra
+    , IsRecentEra (..)
+    , LatestLedgerEra
+    , MaybeInRecentEra (..)
+    , RecentEra (..)
+    , shelleyBasedEra
     )
 import Numeric.Natural
     ( Natural
@@ -283,7 +233,6 @@ import qualified Cardano.Ledger.Alonzo.Core as Alonzo
 import qualified Cardano.Ledger.Alonzo.Scripts as Alonzo
 import qualified Cardano.Ledger.Api as Ledger
 import qualified Cardano.Ledger.Babbage as Babbage
-import qualified Cardano.Ledger.Babbage.Tx as Babbage
 import qualified Cardano.Ledger.Babbage.TxBody as Babbage
 import qualified Cardano.Ledger.Core as Core
 import qualified Cardano.Ledger.Credential as Core
@@ -297,220 +246,6 @@ import qualified Cardano.Wallet.Primitive.Types.Tx.Constraints as W
     ( txOutMaxCoin
     )
 import qualified Data.Map as Map
-import qualified Data.Set as Set
-
---------------------------------------------------------------------------------
--- Eras
---------------------------------------------------------------------------------
-
-type BabbageEra = Ledger.BabbageEra StandardCrypto
-type ConwayEra = Ledger.ConwayEra StandardCrypto
-
-type LatestLedgerEra = StandardConway
-
---------------------------------------------------------------------------------
--- RecentEra
---------------------------------------------------------------------------------
-
--- | 'RecentEra' respresents the eras we care about constructing transactions
--- for.
---
--- To have the same software constructing transactions just before and just
--- after a hard-fork, we need to, at that time, support the two latest eras. We
--- could get away with just supporting one era at other times, but for
--- simplicity we stick with always supporting the two latest eras for now.
---
--- NOTE: We /could/ let 'era' refer to eras from the ledger rather than from
--- cardano-api.
-data RecentEra era where
-    RecentEraBabbage :: RecentEra BabbageEra
-    RecentEraConway :: RecentEra ConwayEra
-
-deriving instance Eq (RecentEra era)
-deriving instance Show (RecentEra era)
-
-instance TestEquality RecentEra where
-    testEquality RecentEraBabbage RecentEraBabbage = Just Refl
-    testEquality RecentEraConway RecentEraConway = Just Refl
-    testEquality RecentEraBabbage RecentEraConway = Nothing
-    testEquality RecentEraConway RecentEraBabbage = Nothing
-
-class
-    ( CardanoApi.IsShelleyBasedEra (CardanoApiEra era)
-    , CardanoApi.ShelleyLedgerEra (CardanoApiEra era) ~ era
-    , Typeable era
-    , RecentEraConstraints era
-    ) => IsRecentEra era where
-    recentEra :: RecentEra era
-
-type family CardanoApiEra era = cardanoApiEra | cardanoApiEra -> era
-type instance CardanoApiEra BabbageEra = CardanoApi.BabbageEra
-type instance CardanoApiEra ConwayEra = CardanoApi.ConwayEra
-
--- | Convenient constraints. Constraints may be dropped as we move to new eras.
---
--- Adding too many constraints shouldn't be a concern as the point of
--- 'RecentEra' is to work with a small closed set of eras, anyway.
-type RecentEraConstraints era =
-    ( Core.Era era
-    , Core.EraTx era
-    , Core.EraCrypto era ~ StandardCrypto
-    , Core.Script era ~ AlonzoScript era
-    , Core.Tx era ~ Babbage.AlonzoTx era
-    , Core.EraTxCert era
-    , Core.Value era ~ Value
-    , Core.TxWits era ~ AlonzoTxWits era
-    , Alonzo.AlonzoEraPParams era
-    , Ledger.AlonzoEraTx era
-    , ScriptsNeeded era ~ AlonzoScriptsNeeded era
-    , AlonzoEraScript era
-    , Eq (TxOut era)
-    , Ledger.Crypto (Core.EraCrypto era)
-    , Show (TxOut era)
-    , Show (Core.Tx era)
-    , Eq (Core.Tx era)
-    , Babbage.BabbageEraTxBody era
-    , Alonzo.AlonzoEraTxBody era
-    , Shelley.EraUTxO era
-    , Show (TxOut era)
-    , Eq (TxOut era)
-    , Show (PParams era)
-    , Show (Script era)
-    , EraPlutusContext era
-    , AllegraEraScript era
-    , Core.NativeScript era ~ Timelock era
-    )
-
--- | Returns a proof that the given era is a recent era.
---
--- Otherwise, returns @Nothing@.
-toRecentEra
-    :: CardanoApi.CardanoEra era
-    -> Maybe (RecentEra (CardanoApi.ShelleyLedgerEra era))
-toRecentEra = \case
-    CardanoApi.ConwayEra  -> Just RecentEraConway
-    CardanoApi.BabbageEra -> Just RecentEraBabbage
-    CardanoApi.AlonzoEra  -> Nothing
-    CardanoApi.MaryEra    -> Nothing
-    CardanoApi.AllegraEra -> Nothing
-    CardanoApi.ShelleyEra -> Nothing
-    CardanoApi.ByronEra   -> Nothing
-
-fromRecentEra :: RecentEra era -> CardanoApi.CardanoEra (CardanoApiEra era)
-fromRecentEra = \case
-    RecentEraConway -> CardanoApi.ConwayEra
-    RecentEraBabbage -> CardanoApi.BabbageEra
-
-instance IsRecentEra BabbageEra where
-    recentEra = RecentEraBabbage
-
-instance IsRecentEra ConwayEra where
-    recentEra = RecentEraConway
-
-cardanoEraFromRecentEra
-    :: RecentEra era
-    -> CardanoApi.CardanoEra (CardanoApiEra era)
-cardanoEraFromRecentEra era = case shelleyBasedEraFromRecentEra era of
-    CardanoApi.ShelleyBasedEraBabbage -> CardanoApi.toCardanoEra CardanoApi.BabbageEra
-    CardanoApi.ShelleyBasedEraConway -> CardanoApi.toCardanoEra CardanoApi.ConwayEra
-    _ -> error "we are expecting only Babbage and Conway"
-
-shelleyBasedEraFromRecentEra
-    :: RecentEra era
-    -> CardanoApi.ShelleyBasedEra (CardanoApiEra era)
-shelleyBasedEraFromRecentEra = \case
-    RecentEraConway -> CardanoApi.ShelleyBasedEraConway
-    RecentEraBabbage -> CardanoApi.ShelleyBasedEraBabbage
-
--- Similar to 'CardanoApi.cardanoEra', but with an 'IsRecentEra era' constraint
--- instead of 'CardanoApi.IsCardanoEra'.
-cardanoEra
-    :: forall era. IsRecentEra era
-    => CardanoApi.CardanoEra (CardanoApiEra era)
-cardanoEra = cardanoEraFromRecentEra $ recentEra @era
-
--- | For convenience working with 'IsRecentEra'.
---
--- Similar to 'CardanoApi.shelleyBasedEra, but with a 'IsRecentEra era'
--- constraint instead of 'CardanoApi.IsShelleyBasedEra'.
-shelleyBasedEra
-    :: forall era. IsRecentEra era
-    => CardanoApi.ShelleyBasedEra (CardanoApiEra era)
-shelleyBasedEra = shelleyBasedEraFromRecentEra $ recentEra @era
-
-data MaybeInRecentEra (thing :: Type -> Type)
-    = InNonRecentEraByron
-    | InNonRecentEraShelley
-    | InNonRecentEraAllegra
-    | InNonRecentEraMary
-    | InNonRecentEraAlonzo
-    | InRecentEraBabbage (thing BabbageEra)
-    | InRecentEraConway (thing ConwayEra)
-
-deriving instance (Eq (a BabbageEra), (Eq (a ConwayEra)))
-    => Eq (MaybeInRecentEra a)
-deriving instance (Show (a BabbageEra), (Show (a ConwayEra)))
-    => Show (MaybeInRecentEra a)
-
--- | An existential type like 'AnyCardanoEra', but for 'RecentEra'.
-data AnyRecentEra where
-    AnyRecentEra
-        :: IsRecentEra era -- Provide class constraint
-        => RecentEra era   -- and explicit value.
-        -> AnyRecentEra    -- and that's it.
-
-instance Enum AnyRecentEra where
-    -- NOTE: We're not starting at 0! 0 would be Byron, which is not a recent
-    -- era.
-    fromEnum = fromEnum . toAnyCardanoEra
-    toEnum n = fromMaybe err . fromAnyCardanoEra $ toEnum n
-      where
-        err = error $ unwords
-            [ "AnyRecentEra.toEnum:", show n
-            , "doesn't correspond to a recent era."
-            ]
-
-instance Bounded AnyRecentEra where
-    minBound = AnyRecentEra RecentEraBabbage
-    maxBound = AnyRecentEra RecentEraConway
-
-instance Ord AnyRecentEra where
-    compare = compare `on` fromEnum
-
-instance Show AnyRecentEra where
-    show (AnyRecentEra era) = "AnyRecentEra " <> show era
-
-instance Eq AnyRecentEra where
-    AnyRecentEra e1 == AnyRecentEra e2 =
-        isJust $ testEquality e1 e2
-
--- | The complete set of recent eras.
---
-allRecentEras :: Set AnyRecentEra
-allRecentEras = Set.fromList [minBound .. maxBound]
-
-toAnyCardanoEra :: AnyRecentEra -> CardanoApi.AnyCardanoEra
-toAnyCardanoEra (AnyRecentEra era) =
-    CardanoApi.AnyCardanoEra (fromRecentEra era)
-
-fromAnyCardanoEra
-    :: CardanoApi.AnyCardanoEra
-    -> Maybe AnyRecentEra
-fromAnyCardanoEra = \case
-    CardanoApi.AnyCardanoEra CardanoApi.ByronEra ->
-        Nothing
-    CardanoApi.AnyCardanoEra CardanoApi.ShelleyEra ->
-        Nothing
-    CardanoApi.AnyCardanoEra CardanoApi.AllegraEra ->
-        Nothing
-    CardanoApi.AnyCardanoEra CardanoApi.MaryEra ->
-        Nothing
-    CardanoApi.AnyCardanoEra CardanoApi.AlonzoEra ->
-        Nothing
-    CardanoApi.AnyCardanoEra CardanoApi.BabbageEra ->
-        Just $ AnyRecentEra RecentEraBabbage
-    CardanoApi.AnyCardanoEra CardanoApi.ConwayEra ->
-        Just $ AnyRecentEra RecentEraConway
 
 --------------------------------------------------------------------------------
 -- Key witness counts
