+++ lib/balance-tx/lib/internal/Internal/Cardano/Write/Tx/TxWithUTxO/Gen.hs
@@ -30,9 +30,11 @@ import Cardano.Ledger.Api.Tx.Body
 import Control.Lens
     ( view
     )
-import Internal.Cardano.Write.Tx
+import Internal.Cardano.Write.Eras
     ( IsRecentEra
-    , Tx
+    )
+import Internal.Cardano.Write.Tx
+    ( Tx
     , TxIn
     , TxOut
     , UTxO (UTxO)
